requirements.txt
astroid==1.6.3
attrs==18.1.0
backcall==0.1.0
bleach==2.1.3
click==6.7
colorama==0.3.9
decorator==4.3.0
dominate==2.3.1
Faker==0.8.13
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Moment==0.6.0
Flask-PageDown==0.2.2
Flask-Script==2.0.6
flask-shell==0.1.3
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
html5lib==1.0.1
ipython==6.3.1
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
parso==0.2.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
py==1.5.3
Pygments==2.2.0
pylint==1.8.4
pytest==3.5.1
pytest-flask==0.10.0
python-dateutil==2.7.2
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.6
text-unidecode==1.2
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
